Fix narrowing conversion warnings.

+++ amr-wind/utilities/sampling/SamplingContainer.cpp
@@ -120,12 +120,12 @@ void SamplingContainer::initialize_particles(
     ptile.resize(num_particles);
 
     int pidx = 0;
-    const int nextid = ParticleType::NextID();
+    const int nextid = static_cast<int>(ParticleType::NextID());
     auto* pstruct = ptile.GetArrayOfStructs()().data();
     SamplerBase::SampleLocType locs;
     for (const auto& probe : samplers) {
         probe->sampling_locations(locs);
-        const int npts = locs.size();
+        const int npts = static_cast<int>(locs.size());
         const auto probe_id = probe->id();
         amrex::Gpu::DeviceVector<amrex::Array<amrex::Real, AMREX_SPACEDIM>>
             dlocs(npts);
@@ -251,7 +251,8 @@ void SamplingContainer::populate_buffer(std::vector<double>& buf)
     amrex::Gpu::copy(
         amrex::Gpu::deviceToHost, dbuf.begin(), dbuf.end(), buf.begin());
     amrex::ParallelDescriptor::ReduceRealSum(
-        buf.data(), buf.size(), amrex::ParallelDescriptor::IOProcessorNumber());
+        buf.data(), static_cast<int>(buf.size()),
+        amrex::ParallelDescriptor::IOProcessorNumber());
 }
 
 } // namespace amr_wind::sampling
